fix: registration photo
