+++ src/handlers/message/__init__.py
@@ -1,1 +1,1 @@
-from . import file, registration
+from . import file, registration, photo
