+++ src/handlers/message/file.py
@@ -7,11 +7,15 @@ from aiogram.types import ContentType
 from starlette_context import context
 from starlette_context.header_keys import HeaderKeys
 
+from src.handlers.states.auth import AuthGroup
 from src.handlers.states.file import FileStates
 from src.logger import logger
 from src.metrics import TOTAL_SEND_MESSAGES, measure_time
 from src.schema.file import FileMessage
-from src.storage.minio_client import check_minio_connection, upload_file
+from src.schema.profile import ProfileCreate
+from src.schema.registration import RegistrationMessage
+from src.schema.user import UserCreate
+from src.storage.minio_client import check_minio_connection, get_file_path, upload_file
 from src.storage.rabbit import channel_pool
 
 from .router import router
@@ -31,57 +35,62 @@ def shorten_file_name(file_name: str, max_length: int = 32) -> str:
 @measure_time('handle_file_upload')
 async def handle_file_upload(message: types.Message, state: FSMContext) -> None:
     """
-    Обрабатывает файл, если состояние пользователя ожидает файл.
+    Обрабатывает файл в двух случаях:
+    1. Когда пользователь в состоянии FileStates.waiting_for_file (команда /upload)
+    2. Когда пользователь в состоянии AuthGroup.registration_photo (процесс регистрации)
     """
     # Проверяем текущее состояние пользователя
     current_state = await state.get_state()
 
-    if current_state == FileStates.waiting_for_file.state:
-        if message.from_user is None:
-            logger.error('Ошибка: сообщение не содержит информации об отправителе (from_user = None).')
-            return
+    # Проверяем только нужные состояния
+    if current_state not in [FileStates.waiting_for_file.state, AuthGroup.registration_photo.state]:
+        await message.reply('Ваш запрос некорректен. Используйте команду для загрузки файла.')
+        return
 
-        # Если ожидается файл
-        if not message.document:
-            await message.reply('Пожалуйста, отправьте файл.')
-            return
+    if message.from_user is None:
+        logger.error('Ошибка: сообщение не содержит информации об отправителе (from_user = None).')
+        return
 
-        # Проверяем соединение с MinIO
-        if not check_minio_connection():
-            await message.reply('Ошибка: сервис хранения файлов недоступен.')
-            return
+    # Проверяем соединение с MinIO
+    if not check_minio_connection():
+        await message.reply('Ошибка: сервис хранения файлов недоступен.')
+        return
 
-        from src.bot import bot
+    from src.bot import bot
 
-        document = message.document
-        file_info = await bot.get_file(document.file_id)
+    if not message.document:
+        await message.reply('Пожалуйста, отправьте файл.')
+        return
 
-        if file_info.file_path is None or document.file_name is None:
-            logger.error('Ошибка: не удалось получить информацию о файле. ID файла: %s', document.file_id)
-            await message.reply('Ошибка при получении информации о файле.')
-            return
+    document = message.document
+    file_info = await bot.get_file(document.file_id)
+    if file_info.file_path is None or document.file_name is None:
+        logger.error('Ошибка: не удалось получить информацию о файле. ID файла: %s', document.file_id)
+        await message.reply('Ошибка при получении информации о файле.')
+        return
 
-        file_bytes = await bot.download_file(file_info.file_path)
+    file_bytes = await bot.download_file(file_info.file_path)
+    if file_bytes is None:
+        logger.error('Ошибка: не удалось скачать файл.')
+        await message.reply('Ошибка при скачивании файла.')
+        return
 
-        if file_bytes is None:
-            logger.error('Ошибка: не удалось скачать файл. ID файла: %s', document.file_id)
-            await message.reply('Ошибка при скачивании файла.')
-            return
+    user_id = message.from_user.id
+    file_name = shorten_file_name(document.file_name)
 
-        user_id = message.from_user.id
-        file_name = shorten_file_name(document.file_name)
+    try:
+        unique_name = upload_file(user_id, file_name, file_bytes.read())
+        logger.info('Файл %s загружен. ID пользователя: %s. Путь к файлу: %s', file_name, user_id, unique_name)
 
-        try:
-            unique_name = upload_file(user_id, file_name, file_bytes.read())
-            logger.info('Файл %s загружен. ID пользователя: %s. Путь к файлу: %s', file_name, user_id, unique_name)
+        # Подключаемся к очереди
+        async with channel_pool.acquire() as channel:
+            # Объявляем обменник и очередь
+            exchange = await channel.declare_exchange('user_files', ExchangeType.TOPIC, durable=True)
+            queue = await channel.declare_queue('user_messages', durable=True)
+            await queue.bind(exchange, 'user_messages')
 
-            # Подключаемся к очереди
-            async with channel_pool.acquire() as channel:
-                # Объявляем обменник и очередь
-                exchange = await channel.declare_exchange('user_files', ExchangeType.TOPIC, durable=True)
-                queue = await channel.declare_queue('user_messages', durable=True)
-                await queue.bind(exchange, 'user_messages')
-
+            if current_state == FileStates.waiting_for_file.state:
+                # Обработка обычной загрузки файла
                 await exchange.publish(
                     aio_pika.Message(
                         msgpack.packb(
@@ -95,14 +104,59 @@ async def handle_file_upload(message: types.Message, state: FSMContext) -> None:
                     ),
                     routing_key='user_messages',
                 )
+                TOTAL_SEND_MESSAGES.labels(operation='upload_file').inc()
+                await state.clear()
+                await message.reply(f'Файл {file_name} успешно загружен!')
 
-            TOTAL_SEND_MESSAGES.labels(operation='upload_file').inc()
+            elif current_state == AuthGroup.registration_photo.state:
+                # Проверяем, что это изображение только для регистрации
+                if not document.mime_type.startswith('image/'):
+                    await message.reply('Пожалуйста, отправьте изображение для профиля.')
+                    return
 
-            # Сбрасываем состояние
-            await state.clear()
-            await message.reply(f'Файл {file_name} успешно загружен!')
-        except Exception as e:
-            logger.error('Ошибка при загрузке файла: %s', e)
-            await message.reply('Произошла ошибка при загрузке файла. Пожалуйста, попробуйте еще раз.')
-    else:
-        await message.reply('Ваш запрос некорректен. Используйте команду для загрузки файла.')
+                # Обработка фото для регистрации
+                photo_url = get_file_path(unique_name)
+                user_data = await state.get_data()
+
+                try:
+                    user = UserCreate(
+                        user_id=user_id,
+                        telegram_id=message.from_user.id,
+                        first_name=user_data['first_name'],
+                        age=user_data['age'],
+                        gender=user_data['gender'],
+                        city=user_data['city'],
+                        bio=user_data['bio'],
+                        photo_url=photo_url,
+                    )
+
+                    profile = ProfileCreate(
+                        user_id=user_id,
+                        bio=user_data['bio'],
+                        photo_url=photo_url,
+                    )
+
+                    await exchange.publish(
+                        aio_pika.Message(
+                            body=msgpack.packb(
+                                RegistrationMessage(
+                                    user=user,
+                                    profile=profile,
+                                    correlation_id=context.get(HeaderKeys.correlation_id)
+                                ).model_dump()
+                            ),
+                            content_type='application/x-msgpack',
+                            headers={'correlation_id': context.get(HeaderKeys.correlation_id)},
+                        ),
+                        routing_key='user_messages',
+                    )
+                    TOTAL_SEND_MESSAGES.labels(operation='registration').inc()
+                    await state.clear()
+                    await message.reply('Спасибо за регистрацию!')
+                except Exception as e:
+                    logger.error('Ошибка при создании пользователя: %s', e)
+                    await message.reply('Произошла ошибка при обработке данных. Пожалуйста, попробуйте еще раз.')
+
+    except Exception as e:
+        logger.error('Ошибка при загрузке файла: %s', e)
+        await message.reply('Произошла ошибка при загрузке файла. Пожалуйста, попробуйте еще раз.')
